refactor(phase8-g1): carve export v1 shim to knowledge_base domain (#47)
Per PM scope (msg=7733c905) and Bryce inventory G1: complete the export
carve trail started by #1670 (ExportTask ORM) and #1672 (export_collection_task).
Pure carve / move; URL contract `/api/v1/export*` and `/api/v1/export-tasks*`
unchanged.

Changes:
- aperag/views/export.py → aperag/domains/knowledge_base/api/export_routes.py
  (3 routes: create_export_task / get_export_task / download_export;
  switched User ORM dep → AuthenticatedUser Protocol, matching the
  canonical KB routes pattern)
- aperag/service/export_service.py → aperag/domains/knowledge_base/service/export_service.py
  (git mv preserves history; replaced view_models.ExportTaskResponse
  references with bare ExportTaskResponse imported from KB schemas)
- ExportTaskResponse schema migrated from aperag/schema/view_models.py
  to aperag/domains/knowledge_base/schemas.py (added to __all__);
  view_models.py keeps a backward-compat re-export following the same
  pattern used for knowledge_graph/retrieval schemas
- aperag/app.py: import switched from aperag.views.export to
  aperag.domains.knowledge_base.api.export_routes; mount prefix
  remains /api/v1 (D1 hard-cut to /api/v2 deferred to later G3+ batch)

Verification:
- pytest tests/unit_test/test_modularization_boundaries.py -x → 21 passed
  (G1-G19 all green; new KB route does not import legacy aggregates)
- pytest tests/unit_test/ → 686 passed, 29 skipped (matches phase8 baseline)
- grep "from aperag.service.export_service|from aperag.views.export"
  returns 0 hits in aperag/, tests/, config/
- python -c "from aperag.app import app; ..." confirms 3 export endpoints
  remain at /api/v1/collections/{id}/export, /api/v1/export-tasks/{id},
  /api/v1/export-tasks/{id}/download

aperag/service/ retains exactly the 3 permanent seam files
(quota_service / prompt_template_service / search_pipeline_service);
no Layer A / Layer B violation.

Ghost-check: none.
